% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{mizerExamples}
\alias{mizerExamples}
\title{mizerExamples: Example calibrations for Mizer}
\description{
The mizerExamples package contains several example calibrations to be run with \pkg{mizer}.
}
\details{
Using \pkg{mizerExamples} is relatively simple: load the parameters and simulate.

The package contains the following calibrations:
\link{NorthSea_params},
\link{Baltic_params},
\link{Benguela_params},
\link{NEUSCS_params}, and
\link{Barents_params}.

See the help pages for each calibration for full description.
}
\examples{
param = NorthSea_params
sim = project(param)
plot(sim)

}
